storage: scope schema-change hints to dropped constraints
Move SchemaChangeError and its variants into their own module, include the
table oid in the diagnosis, and only attach recovery hints to the cases the
EXCLUDE CONSTRAINTS work is about: a dropped, renamed, or recreated
PRIMARY KEY/UNIQUE constraint and a dropped NOT NULL constraint. Column
drops, type changes, position changes, and table renames keep their
specific message but carry no hint.

Testdrive asserts the exact hint text for a dropped UNIQUE constraint via
the table's status details.
