+++ test/pg-cdc-old-syntax/alter-table-after-source.td
@@ -143,8 +143,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE remove_column DROP COLUMN f2;
 
 ! SELECT * from remove_column;
-contains:incompatible schema change on public.remove_column: column "f2" was dropped or renamed upstream
-hint:WITH (EXCLUDE COLUMNS ("f2")) before the upstream drop
+regex:incompatible schema change on public\.remove_column \(oid \d+\): column "f2" was dropped or renamed upstream
 
 
 #
@@ -157,8 +156,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_column ALTER COLUMN f2 TYPE CHAR(2);
 
 ! SELECT * from alter_column;
-contains:incompatible schema change on public.alter_column: the type of column "f2" changed upstream
-hint:WITH (TEXT COLUMNS ("f2"))
+regex:incompatible schema change on public\.alter_column \(oid \d+\): the type of column "f2" changed upstream
 
 
 #
@@ -171,7 +169,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_drop_nullability ALTER COLUMN f1 DROP NOT NULL;
 
 ! SELECT * FROM alter_drop_nullability WHERE f1 IS NOT NULL;
-contains:incompatible schema change on public.alter_drop_nullability: the NOT NULL constraint on column "f1" was dropped upstream
+regex:incompatible schema change on public\.alter_drop_nullability \(oid \d+\): the NOT NULL constraint on column "f1" was dropped upstream
 hint:WITH (EXCLUDE ALL CONSTRAINTS)
 
 # We have guaranteed that this column is not null so the optimizer eagerly
@@ -204,7 +202,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_drop_pk DROP CONSTRAINT alter_drop_pk_pkey;
 
 ! SELECT f1 FROM alter_drop_pk;
-contains:incompatible schema change on public.alter_drop_pk: PRIMARY KEY constraint "alter_drop_pk_pkey" (f1) was dropped upstream
+regex:incompatible schema change on public\.alter_drop_pk \(oid \d+\): PRIMARY KEY constraint "alter_drop_pk_pkey" \(f1\) was dropped upstream
 hint:WITH (EXCLUDE CONSTRAINTS ('alter_drop_pk_pkey')) before the upstream drop
 
 
@@ -234,7 +232,7 @@ ALTER TABLE alter_cycle_pk DROP CONSTRAINT alter_cycle_pk_pkey;
 ALTER TABLE alter_cycle_pk ADD PRIMARY KEY(f1);
 
 ! SELECT * FROM alter_cycle_pk;
-contains:incompatible schema change on public.alter_cycle_pk: PRIMARY KEY constraint "alter_cycle_pk_pkey" (f1) was dropped
+regex:incompatible schema change on public\.alter_cycle_pk \(oid \d+\): PRIMARY KEY constraint "alter_cycle_pk_pkey" \(f1\) was renamed or recreated upstream
 hint:EXCLUDE CONSTRAINTS ('alter_cycle_pk_pkey')
 
 
@@ -265,8 +263,13 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_drop_unique DROP CONSTRAINT alter_drop_unique_f1_key;
 
 ! SELECT f1 FROM alter_drop_unique;
-contains:incompatible schema change on public.alter_drop_unique: UNIQUE constraint "alter_drop_unique_f1_key" (f1) was dropped upstream
+regex:incompatible schema change on public\.alter_drop_unique \(oid \d+\): UNIQUE constraint "alter_drop_unique_f1_key" \(f1\) was dropped upstream
 hint:WITH (EXCLUDE CONSTRAINTS ('alter_drop_unique_f1_key')) before the upstream drop
+
+# The same hint is surfaced in the table's status.
+> SELECT details->'hints'->>0 = E'To keep ingesting without this constraint, recreate the table in a new versioned schema, then swap your views to the new table:\n  CREATE SCHEMA v2;\n  CREATE TABLE v2.alter_drop_unique\n  FROM SOURCE <source> (REFERENCE public.alter_drop_unique);\nTo make a planned constraint drop a non-event, create the table with WITH (EXCLUDE CONSTRAINTS (''alter_drop_unique_f1_key'')) before the upstream drop.'
+  FROM mz_internal.mz_source_statuses WHERE name = 'alter_drop_unique'
+true
 
 
 #
@@ -294,8 +297,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_extend_column ALTER COLUMN f1 TYPE VARCHAR(20);
 
 ! SELECT * FROM alter_extend_column;
-contains:incompatible schema change on public.alter_extend_column: the type of column "f1" changed upstream
-hint:WITH (TEXT COLUMNS ("f1"))
+regex:incompatible schema change on public\.alter_extend_column \(oid \d+\): the type of column "f1" changed upstream
 
 
 #
@@ -307,8 +309,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_decimal ALTER COLUMN f1 TYPE DECIMAL(6,1);
 
 ! SELECT * FROM alter_decimal;
-contains:incompatible schema change on public.alter_decimal: the type of column "f1" changed upstream
-hint:WITH (TEXT COLUMNS ("f1"))
+regex:incompatible schema change on public\.alter_decimal \(oid \d+\): the type of column "f1" changed upstream
 
 
 #
@@ -321,8 +322,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_table_rename RENAME TO alter_table_renamed;
 
 ! SELECT * FROM alter_table_rename;
-contains:incompatible schema change on public.alter_table_rename: table was renamed or moved upstream (it is now public.alter_table_renamed with oid
-hint:FROM SOURCE <source> (REFERENCE public.alter_table_renamed)
+regex:incompatible schema change on public\.alter_table_rename \(oid \d+\): table was renamed or moved upstream \(it is now public\.alter_table_renamed with oid
 
 #
 # Alter table rename column
@@ -336,8 +336,7 @@ ALTER TABLE alter_table_rename_column RENAME COLUMN f2 TO f1;
 ALTER TABLE alter_table_rename_column RENAME COLUMN f3 TO f2;
 
 ! SELECT * FROM alter_table_rename_column;
-contains:incompatible schema change on public.alter_table_rename_column: column
-hint:CREATE SCHEMA v2;
+regex:incompatible schema change on public\.alter_table_rename_column \(oid \d+\): column
 
 
 #
@@ -352,8 +351,7 @@ ALTER TABLE alter_table_change_attnum DROP COLUMN f2;
 ALTER TABLE alter_table_change_attnum ADD COLUMN f2 VARCHAR(10);
 
 ! SELECT * FROM alter_table_change_attnum;
-contains:incompatible schema change on public.alter_table_change_attnum: column "f2" changed position upstream
-hint:CREATE SCHEMA v2;
+regex:incompatible schema change on public\.alter_table_change_attnum \(oid \d+\): column "f2" changed position upstream
 
 > SELECT * from alter_table_supported;
 1 1
@@ -379,8 +377,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_table_supported DROP COLUMN f2;
 
 ! SELECT * from alter_table_supported;
-contains:incompatible schema change on public.alter_table_supported: column "f2" was dropped or renamed upstream
-hint:WITH (EXCLUDE COLUMNS ("f2")) before the upstream drop
+regex:incompatible schema change on public\.alter_table_supported \(oid \d+\): column "f2" was dropped or renamed upstream
 
 
 #
