+++ test/pg-cdc/alter-table-after-source-1.td
@@ -180,8 +180,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_column ALTER COLUMN f2 TYPE CHAR(2);
 
 ! SELECT * from alter_column;
-contains:incompatible schema change on public.alter_column: the type of column "f2" changed upstream
-hint:WITH (TEXT COLUMNS ("f2"))
+regex:incompatible schema change on public\.alter_column \(oid \d+\): the type of column "f2" changed upstream
 
 
 #
@@ -194,7 +193,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_drop_nullability ALTER COLUMN f1 DROP NOT NULL;
 
 ! SELECT * FROM alter_drop_nullability WHERE f1 IS NOT NULL;
-contains:incompatible schema change on public.alter_drop_nullability: the NOT NULL constraint on column "f1" was dropped upstream
+regex:incompatible schema change on public\.alter_drop_nullability \(oid \d+\): the NOT NULL constraint on column "f1" was dropped upstream
 hint:WITH (EXCLUDE ALL CONSTRAINTS)
 
 # We have guaranteed that this column is not null so the optimizer eagerly
@@ -227,7 +226,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_drop_pk DROP CONSTRAINT alter_drop_pk_pkey;
 
 ! SELECT f1 FROM alter_drop_pk;
-contains:incompatible schema change on public.alter_drop_pk: PRIMARY KEY constraint "alter_drop_pk_pkey" (f1) was dropped upstream
+regex:incompatible schema change on public\.alter_drop_pk \(oid \d+\): PRIMARY KEY constraint "alter_drop_pk_pkey" \(f1\) was dropped upstream
 hint:WITH (EXCLUDE CONSTRAINTS ('alter_drop_pk_pkey')) before the upstream drop
 
 
@@ -257,7 +256,7 @@ ALTER TABLE alter_cycle_pk DROP CONSTRAINT alter_cycle_pk_pkey;
 ALTER TABLE alter_cycle_pk ADD PRIMARY KEY(f1);
 
 ! SELECT * FROM alter_cycle_pk;
-contains:incompatible schema change on public.alter_cycle_pk: PRIMARY KEY constraint "alter_cycle_pk_pkey" (f1) was dropped
+regex:incompatible schema change on public\.alter_cycle_pk \(oid \d+\): PRIMARY KEY constraint "alter_cycle_pk_pkey" \(f1\) was renamed or recreated upstream
 hint:EXCLUDE CONSTRAINTS ('alter_cycle_pk_pkey')
 
 
@@ -288,11 +287,11 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_drop_unique DROP CONSTRAINT alter_drop_unique_f1_key;
 
 ! SELECT f1 FROM alter_drop_unique;
-contains:incompatible schema change on public.alter_drop_unique: UNIQUE constraint "alter_drop_unique_f1_key" (f1) was dropped upstream
+regex:incompatible schema change on public\.alter_drop_unique \(oid \d+\): UNIQUE constraint "alter_drop_unique_f1_key" \(f1\) was dropped upstream
 hint:WITH (EXCLUDE CONSTRAINTS ('alter_drop_unique_f1_key')) before the upstream drop
 
 # The same hint is surfaced in the table's status.
-> SELECT details->>'hints' LIKE '%EXCLUDE CONSTRAINTS (''alter_drop_unique_f1_key'')%'
+> SELECT details->'hints'->>0 = E'To keep ingesting without this constraint, recreate the table in a new versioned schema, then swap your views to the new table:\n  CREATE SCHEMA v2;\n  CREATE TABLE v2.alter_drop_unique\n  FROM SOURCE <source> (REFERENCE public.alter_drop_unique);\nTo make a planned constraint drop a non-event, create the table with WITH (EXCLUDE CONSTRAINTS (''alter_drop_unique_f1_key'')) before the upstream drop.'
   FROM mz_internal.mz_source_statuses WHERE name = 'alter_drop_unique'
 true
 
@@ -307,8 +306,8 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_rename_unique RENAME CONSTRAINT alter_rename_unique_f1_key TO alter_rename_unique_f1_key_v2;
 
 ! SELECT f1 FROM alter_rename_unique;
-contains:incompatible schema change on public.alter_rename_unique: UNIQUE constraint "alter_rename_unique_f1_key" (f1) was renamed upstream to "alter_rename_unique_f1_key_v2"
-hint:WITH (EXCLUDE CONSTRAINTS ('alter_rename_unique_f1_key_v2'));
+regex:incompatible schema change on public\.alter_rename_unique \(oid \d+\): UNIQUE constraint "alter_rename_unique_f1_key" \(f1\) was renamed or recreated upstream
+hint:WITH (EXCLUDE CONSTRAINTS ('alter_rename_unique_f1_key')) before the upstream drop
 
 
 #
@@ -336,8 +335,7 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_extend_column ALTER COLUMN f1 TYPE VARCHAR(20);
 
 ! SELECT * FROM alter_extend_column;
-contains:incompatible schema change on public.alter_extend_column: the type of column "f1" changed upstream
-hint:WITH (TEXT COLUMNS ("f1"))
+regex:incompatible schema change on public\.alter_extend_column \(oid \d+\): the type of column "f1" changed upstream
 
 
 #
@@ -349,5 +347,4 @@ $ postgres-execute connection=postgres://postgres:postgres@postgres
 ALTER TABLE alter_decimal ALTER COLUMN f1 TYPE DECIMAL(6,1);
 
 ! SELECT * FROM alter_decimal;
-contains:incompatible schema change on public.alter_decimal: the type of column "f1" changed upstream
-hint:WITH (TEXT COLUMNS ("f1"))
+regex:incompatible schema change on public\.alter_decimal \(oid \d+\): the type of column "f1" changed upstream
